do_mounts: always prefer tmpfs for rootfs when available
Inspired by the situation from [0].

The roots of choosing tmpfs/ramfs backend for rootfs go far back
in history, and it's unclear at all why it was decided to select
full-blown tmpfs when "root=" is not specified and feature-poor
ramfs otherwise.
There are several cases when "root=" is not needed at all to work,
and it doesn't break anything or make any [negative] sense. On the
other hand, such separation is rather counter-intuitive and makes
debugging more difficult.
Simply always use tmpfs when it's available -- just like devtmpfs
does [for over a decade].

[0] https://lore.kernel.org/kernel-hardening/20210701234807.50453-1-alobakin@pm.me/

Signed-off-by: Alexander Lobakin <alobakin@pm.me>