treewide: switch most of the low-timeout users to schedule_hrtimeout
Make use of the hrtimeout API in most of the places where the argument of
different colors of schedule_timeout() is or could be lower than 50 ms.
When the caller specifies timeout in jiffies, not converts it from a
fixed value, take 1 ms as a jiffy.
Due the macro magic introduced in the previous commit, most of cases when
the timeout exceeds 4 jiffies will be optimized out back to the regular
functions (to avoid jumping back and forth between hr <-> tick).

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>